+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/handler/ContractRejectionHandler.java
@@ -8,26 +8,23 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
+ *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation, refactoring
  *
  */
 
 package org.eclipse.dataspaceconnector.ids.api.multipart.handler;
 
 import de.fraunhofer.iais.eis.ContractRejectionMessage;
-import de.fraunhofer.iais.eis.Message;
 import org.eclipse.dataspaceconnector.ids.api.multipart.message.MultipartRequest;
 import org.eclipse.dataspaceconnector.ids.api.multipart.message.MultipartResponse;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.ConsumerContractNegotiationManager;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.ProviderContractNegotiationManager;
-import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-import java.util.Objects;
-
-import static org.eclipse.dataspaceconnector.ids.api.multipart.util.RejectionMessageUtil.badParameters;
+import static org.eclipse.dataspaceconnector.ids.api.multipart.util.ResponseUtil.badParameters;
+import static org.eclipse.dataspaceconnector.ids.api.multipart.util.ResponseUtil.createMultipartResponse;
+import static org.eclipse.dataspaceconnector.ids.api.multipart.util.ResponseUtil.processedFromStatusResult;
 
 /**
  * This class handles and processes incoming IDS {@link ContractRejectionMessage}s.
@@ -44,26 +41,22 @@ public class ContractRejectionHandler implements Handler {
             @NotNull String connectorId,
             @NotNull ProviderContractNegotiationManager providerNegotiationManager,
             @NotNull ConsumerContractNegotiationManager consumerNegotiationManager) {
-        this.monitor = Objects.requireNonNull(monitor);
-        this.connectorId = Objects.requireNonNull(connectorId);
-        this.providerNegotiationManager = Objects.requireNonNull(providerNegotiationManager);
-        this.consumerNegotiationManager = Objects.requireNonNull(consumerNegotiationManager);
+        this.monitor = monitor;
+        this.connectorId = connectorId;
+        this.providerNegotiationManager = providerNegotiationManager;
+        this.consumerNegotiationManager = consumerNegotiationManager;
     }
 
     @Override
     public boolean canHandle(@NotNull MultipartRequest multipartRequest) {
-        Objects.requireNonNull(multipartRequest);
-
         return multipartRequest.getHeader() instanceof ContractRejectionMessage;
     }
 
     @Override
-    public @Nullable MultipartResponse handleRequest(@NotNull MultipartRequest multipartRequest, @NotNull ClaimToken claimToken) {
-        Objects.requireNonNull(multipartRequest);
-        Objects.requireNonNull(claimToken);
-
+    public @NotNull MultipartResponse handleRequest(@NotNull MultipartRequest multipartRequest) {
+        var claimToken = multipartRequest.getClaimToken();
         var message = (ContractRejectionMessage) multipartRequest.getHeader();
-        var correlationMessageId = message.getCorrelationMessage(); // TODO correlation msg missing
+        var correlationMessageId = message.getCorrelationMessage();
         var correlationId = message.getTransferContract();
         var rejectionReason = message.getContractRejectionReason();
         monitor.debug(String.format("ContractRejectionHandler: Received contract rejection to " +
@@ -71,27 +64,16 @@ public class ContractRejectionHandler implements Handler {
                 correlationId, rejectionReason));
 
         if (correlationId == null) {
-            return createBadParametersErrorMultipartResponse(message);
+            return createMultipartResponse(badParameters(message, connectorId));
         }
 
         // abort negotiation process (one of them can handle this process by id)
-        var result = providerNegotiationManager.declined(claimToken, String.valueOf(correlationId));
-        if (result.fatalError()) {
-            result = consumerNegotiationManager.declined(claimToken, String.valueOf(correlationId));
+        var negotiationDeclineResult = providerNegotiationManager.declined(claimToken, String.valueOf(correlationId));
+        if (negotiationDeclineResult.fatalError()) {
+            negotiationDeclineResult = consumerNegotiationManager.declined(claimToken, String.valueOf(correlationId));
         }
-
-        if (result.fatalError()) {
-            monitor.debug("ContractRejectionHandler: Could not process contract rejection");
-        }
-
-        return MultipartResponse.Builder.newInstance()
-                .header(ResponseMessageUtil.createMessageProcessedNotificationMessage(connectorId, message))
-                .build();
+    
+        return createMultipartResponse(processedFromStatusResult(negotiationDeclineResult, message, connectorId));
     }
 
-    private MultipartResponse createBadParametersErrorMultipartResponse(Message message) {
-        return MultipartResponse.Builder.newInstance()
-                .header(badParameters(message, connectorId))
-                .build();
-    }
 }