+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/handler/ContractRequestHandler.java
@@ -8,7 +8,7 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
+ *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation, refactoring
  *       Daimler TSS GmbH - introduce factory to create RequestInProcessMessage
  *
  */
@@ -18,30 +18,25 @@ package org.eclipse.dataspaceconnector.ids.api.multipart.handler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.fraunhofer.iais.eis.ContractRequest;
 import de.fraunhofer.iais.eis.ContractRequestMessage;
-import de.fraunhofer.iais.eis.Message;
-import de.fraunhofer.iais.eis.RequestInProcessMessage;
 import org.eclipse.dataspaceconnector.ids.api.multipart.message.MultipartRequest;
 import org.eclipse.dataspaceconnector.ids.api.multipart.message.MultipartResponse;
-import org.eclipse.dataspaceconnector.ids.api.multipart.message.ids.IdsResponseMessageFactory;
-import org.eclipse.dataspaceconnector.ids.api.multipart.message.ids.exceptions.InvalidCorrelationMessageException;
 import org.eclipse.dataspaceconnector.ids.spi.IdsIdParser;
 import org.eclipse.dataspaceconnector.ids.spi.Protocols;
 import org.eclipse.dataspaceconnector.ids.spi.transform.ContractTransformerInput;
 import org.eclipse.dataspaceconnector.ids.spi.transform.IdsTransformerRegistry;
 import org.eclipse.dataspaceconnector.spi.asset.AssetIndex;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.ProviderContractNegotiationManager;
-import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractOfferRequest;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractOffer;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
-import java.util.Objects;
 
-import static org.eclipse.dataspaceconnector.ids.api.multipart.util.RejectionMessageUtil.badParameters;
+import static org.eclipse.dataspaceconnector.ids.api.multipart.util.ResponseUtil.badParameters;
+import static org.eclipse.dataspaceconnector.ids.api.multipart.util.ResponseUtil.createMultipartResponse;
+import static org.eclipse.dataspaceconnector.ids.api.multipart.util.ResponseUtil.inProcessFromStatusResult;
 import static org.eclipse.dataspaceconnector.ids.spi.IdsConstants.IDS_WEBHOOK_ADDRESS_PROPERTY;
 
 /**
@@ -55,37 +50,30 @@ public class ContractRequestHandler implements Handler {
     private final ProviderContractNegotiationManager negotiationManager;
     private final IdsTransformerRegistry transformerRegistry;
     private final AssetIndex assetIndex;
-    private final IdsResponseMessageFactory responseMessageFactory;
 
     public ContractRequestHandler(
             @NotNull Monitor monitor,
             @NotNull String connectorId,
             @NotNull ObjectMapper objectMapper,
             @NotNull ProviderContractNegotiationManager negotiationManager,
-            @NotNull IdsResponseMessageFactory responseMessageFactory,
             @NotNull IdsTransformerRegistry transformerRegistry,
             @NotNull AssetIndex assetIndex) {
-        this.monitor = Objects.requireNonNull(monitor);
-        this.connectorId = Objects.requireNonNull(connectorId);
-        this.objectMapper = Objects.requireNonNull(objectMapper);
-        this.negotiationManager = Objects.requireNonNull(negotiationManager);
-        this.transformerRegistry = Objects.requireNonNull(transformerRegistry);
-        this.assetIndex = Objects.requireNonNull(assetIndex);
-        this.responseMessageFactory = Objects.requireNonNull(responseMessageFactory);
+        this.monitor = monitor;
+        this.connectorId = connectorId;
+        this.objectMapper = objectMapper;
+        this.negotiationManager = negotiationManager;
+        this.transformerRegistry = transformerRegistry;
+        this.assetIndex = assetIndex;
     }
 
     @Override
     public boolean canHandle(@NotNull MultipartRequest multipartRequest) {
-        Objects.requireNonNull(multipartRequest);
-
         return multipartRequest.getHeader() instanceof ContractRequestMessage;
     }
 
     @Override
-    public @Nullable MultipartResponse handleRequest(@NotNull MultipartRequest multipartRequest, @NotNull ClaimToken claimToken) {
-        Objects.requireNonNull(multipartRequest);
-        Objects.requireNonNull(claimToken);
-
+    public @NotNull MultipartResponse handleRequest(@NotNull MultipartRequest multipartRequest) {
+        var claimToken = multipartRequest.getClaimToken();
         var message = (ContractRequestMessage) multipartRequest.getHeader();
 
         ContractRequest contractRequest;
@@ -93,27 +81,28 @@ public class ContractRequestHandler implements Handler {
             contractRequest = objectMapper.readValue(multipartRequest.getPayload(), ContractRequest.class);
         } catch (IOException e) {
             monitor.severe("ContractRequestHandler: Contract Request is invalid", e);
-            return createBadParametersErrorMultipartResponse(message);
+            return createMultipartResponse(badParameters(message, connectorId));
         }
-
+    
+        // Get webhook address of requesting connector from message properties
         var idsWebhookAddress = message.getProperties().get(IDS_WEBHOOK_ADDRESS_PROPERTY);
         if (idsWebhookAddress == null || idsWebhookAddress.toString().isBlank()) {
             var msg = "Ids webhook address is invalid";
             monitor.debug(String.format("ContractRequestHandler: %s", msg));
-            return createBadParametersErrorMultipartResponse(message, msg);
+            return createMultipartResponse(badParameters(message, connectorId), msg);
         }
 
         // extract target from contract request
         var permission = contractRequest.getPermission().stream().findFirst().orElse(null);
         if (permission == null) {
             monitor.debug("ContractRequestHandler: Contract Request is invalid");
-            return createBadParametersErrorMultipartResponse(message);
+            return createMultipartResponse(badParameters(message, connectorId));
         }
 
         var target = permission.getTarget();
         if (target == null || String.valueOf(target).isBlank()) {
             monitor.debug("ContractRequestHandler: Contract Request is invalid");
-            return createBadParametersErrorMultipartResponse(message);
+            return createMultipartResponse(badParameters(message, connectorId));
         }
 
         // search for matching asset
@@ -122,7 +111,7 @@ public class ContractRequestHandler implements Handler {
         if (asset == null) {
             var msg = "Target id is invalid";
             monitor.debug(String.format("ContractRequestHandler: %s", msg));
-            return createBadParametersErrorMultipartResponse(message, msg);
+            return createMultipartResponse(badParameters(message, connectorId), msg);
         }
 
         // Create contract offer request
@@ -135,7 +124,7 @@ public class ContractRequestHandler implements Handler {
         if (result.failed()) {
             monitor.debug(String.format("Could not transform contract request: [%s]",
                     String.join(", ", result.getFailureMessages())));
-            return createBadParametersErrorMultipartResponse(message);
+            return createMultipartResponse(badParameters(message, connectorId));
         }
 
         var contractOffer = result.getContent();
@@ -147,41 +136,11 @@ public class ContractRequestHandler implements Handler {
                 .correlationId(String.valueOf(message.getTransferContract()))
                 .contractOffer(contractOffer)
                 .build();
-
-        Message response;
-
-        try {
-            response = responseMessageFactory.createRequestInProcessMessage(message);
-        } catch (Exception e) {
-            if (e instanceof InvalidCorrelationMessageException) {
-                monitor.debug(String.format("Rejecting invalid IDS contract request message [Msg-ID: %s]", message.getId()), e);
-            } else {
-                monitor.severe(String.format("Exception while creating IDS RequestInProcessMessage to answer contract request [Msg-ID: %s]", message.getId()), e);
-            }
-
-            response = responseMessageFactory.createRejectionMessage(message, e);
-        }
-
-        if (response instanceof RequestInProcessMessage) {
-            // Start negotiation process
-            negotiationManager.requested(claimToken, requestObj);
-        }
-
-        return MultipartResponse.Builder.newInstance()
-                .header(response)
-                .build();
-    }
-
-    private MultipartResponse createBadParametersErrorMultipartResponse(Message message) {
-        return MultipartResponse.Builder.newInstance()
-                .header(badParameters(message, connectorId))
-                .build();
-    }
-
-    private MultipartResponse createBadParametersErrorMultipartResponse(Message message, String payload) {
-        return MultipartResponse.Builder.newInstance()
-                .header(badParameters(message, connectorId))
-                .payload(payload)
-                .build();
+    
+        //TODO use ContractNegotiationService for initiation after project structure review
+        
+        var negotiationInitiateResult = negotiationManager.requested(claimToken, requestObj);
+    
+        return createMultipartResponse(inProcessFromStatusResult(negotiationInitiateResult, message, connectorId));
     }
 }