+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/message/MultipartResponse.java
@@ -20,13 +20,16 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.Objects;
 
+/**
+ * Represents an IDS multipart response. Contains the IDS message header and an optional payload.
+ */
 public class MultipartResponse {
 
     private final Message header;
     private final Object payload;
 
     private MultipartResponse(@NotNull Message header, @Nullable Object payload) {
-        this.header = Objects.requireNonNull(header);
+        this.header = header;
         this.payload = payload;
     }
 
@@ -63,6 +66,7 @@ public class MultipartResponse {
         }
 
         public MultipartResponse build() {
+            Objects.requireNonNull(header, "Multipart response header is null.");
             return new MultipartResponse(header, payload);
         }
     }