+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/test/java/org/eclipse/dataspaceconnector/ids/api/multipart/AbstractMultipartControllerIntegrationTest.java
@@ -20,8 +20,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import de.fraunhofer.iais.eis.Contract;
 import de.fraunhofer.iais.eis.ContractAgreementMessage;
 import de.fraunhofer.iais.eis.ContractAgreementMessageBuilder;
-import de.fraunhofer.iais.eis.ContractOfferMessage;
-import de.fraunhofer.iais.eis.ContractOfferMessageBuilder;
 import de.fraunhofer.iais.eis.ContractRejectionMessage;
 import de.fraunhofer.iais.eis.ContractRejectionMessageBuilder;
 import de.fraunhofer.iais.eis.ContractRequestMessage;
@@ -139,7 +137,8 @@ abstract class AbstractMultipartControllerIntegrationTest {
     protected DescriptionRequestMessage getDescriptionRequestMessage(IdsId idsId) {
         DescriptionRequestMessageBuilder builder = new DescriptionRequestMessageBuilder()
                 ._securityToken_(getDynamicAttributeToken())
-                ._issuerConnector_(URI.create("issuerConnector"));
+                ._issuerConnector_(URI.create("issuerConnector"))
+                ._senderAgent_(URI.create("senderAgent"));
 
         if (idsId != null) {
             builder._requestedElement_(
@@ -152,7 +151,7 @@ abstract class AbstractMultipartControllerIntegrationTest {
         var message = new ContractRequestMessageBuilder()
                 ._correlationMessage_(URI.create("correlationId"))
                 ._securityToken_(getDynamicAttributeToken())
-                ._senderAgent_(URI.create("sender"))
+                ._senderAgent_(URI.create("senderAgent"))
                 ._issuerConnector_(URI.create("issuerConnector"))
                 .build();
         message.setProperty(IDS_WEBHOOK_ADDRESS_PROPERTY, "http://someUrl");
@@ -164,6 +163,7 @@ abstract class AbstractMultipartControllerIntegrationTest {
                 ._correlationMessage_(URI.create("correlationId"))
                 ._securityToken_(getDynamicAttributeToken())
                 ._issuerConnector_(URI.create("issuerConnector"))
+                ._senderAgent_(URI.create("senderAgent"))
                 .build();
     }
 
@@ -173,15 +173,7 @@ abstract class AbstractMultipartControllerIntegrationTest {
                 ._transferContract_(URI.create("contractId"))
                 ._securityToken_(getDynamicAttributeToken())
                 ._issuerConnector_(URI.create("issuerConnector"))
-                .build();
-    }
-
-    protected ContractOfferMessage getContractOfferMessage() {
-        return new ContractOfferMessageBuilder()
-                ._correlationMessage_(URI.create("correlationId"))
-                ._senderAgent_(URI.create("sender"))
-                ._securityToken_(getDynamicAttributeToken())
-                ._issuerConnector_(URI.create("issuerConnector"))
+                ._senderAgent_(URI.create("senderAgent"))
                 .build();
     }